break up backend into separate files; inject taxiiStorer; update tests
