+++ taxii_api_root.go
@@ -9,7 +9,7 @@ import (
 
 /* handler */
 
-func trimSlahes(s string) string {
+func trimSlashes(s string) string {
 	re := regexp.MustCompile("^/")
 	s = re.ReplaceAllString(s, "")
 
@@ -20,28 +20,27 @@ func trimSlahes(s string) string {
 	return strings.Join(parts, "/")
 }
 
-func handleTaxiiAPIRoot(w http.ResponseWriter, r *http.Request) {
-	defer recoverFromPanic(w)
-	info.Println("API Root requested for", r.URL)
+func handleTaxiiAPIRoot(ts taxiiStorer) http.HandlerFunc {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		defer recoverFromPanic(w)
+		info.Println("API Root requested for", r.URL)
 
-	ta := taxiiAPIRoot{}
+		ta := taxiiAPIRoot{}
+		err := ta.read(ts, trimSlashes(r.URL.Path))
+		if err != nil {
+			badRequest(w, err)
+			return
+		}
 
-	info.Println("reading path:", trimSlahes(r.URL.Path))
-
-	err := ta.read(trimSlahes(r.URL.Path))
-	if err != nil {
-		badRequest(w, err)
-		return
-	}
-
-	if ta.Title == "" {
-		resourceNotFound(w, errors.New("API Root not defined"))
-	} else {
-		writeContent(w, taxiiContentType, resourceToJSON(ta))
-	}
+		if ta.Title == "" {
+			resourceNotFound(w, errors.New("API Root not defined"))
+		} else {
+			writeContent(w, taxiiContentType, resourceToJSON(ta))
+		}
+	})
 }
 
-/* model */
+/* models */
 
 type taxiiAPIRoot struct {
 	Title            string   `json:"title"`
@@ -50,26 +49,43 @@ type taxiiAPIRoot struct {
 	MaxContentLength int64    `json:"max_content_length"`
 }
 
-func (ta *taxiiAPIRoot) create(path string) error {
+func (ta *taxiiAPIRoot) create(ts taxiiStorer, path string) error {
 	id, err := newTaxiiID()
 	if err != nil {
 		return err
 	}
 
-	err = createResource("taxiiAPIRoot",
+	err = createResource(ts, "taxiiAPIRoot",
 		[]interface{}{id, path, ta.Title, ta.Description, strings.Join(ta.Versions, ","), ta.MaxContentLength})
 	return err
 }
 
-func (ta *taxiiAPIRoot) read(path string) error {
+func (ta *taxiiAPIRoot) read(ts taxiiStorer, path string) error {
 	apiRoot := *ta
 
-	result, err := readResource("taxiiAPIRoot", []interface{}{path})
+	result, err := ts.read("taxiiAPIRoot", []interface{}{path})
 	if err != nil {
 		return err
 	}
 	apiRoot = result.(taxiiAPIRoot)
 
 	*ta = apiRoot
+	return err
+}
+
+type taxiiAPIRoots struct {
+	RootPaths []string
+}
+
+func (ta *taxiiAPIRoots) read(ts taxiiStorer) error {
+	roots := *ta
+
+	result, err := ts.read("taxiiAPIRoots", []interface{}{})
+	if err != nil {
+		return err
+	}
+	roots = result.(taxiiAPIRoots)
+
+	*ta = roots
 	return err
 }
