+++ taxii_api_root_test.go
@@ -6,14 +6,20 @@ import (
 )
 
 func TestHandleTaxiiAPIRoot(t *testing.T) {
-	status, result := handlerTest(handleTaxiiAPIRoot, "GET", apiRootURL, nil)
+	ts, err := newTaxiiStorer(config.DataStore["name"], config.DataStore["path"])
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer ts.disconnect()
+
+	status, result := handlerTest(handleTaxiiAPIRoot(ts), "GET", testAPIRootURL, nil)
 
 	if status != 200 {
 		t.Error("Got:", status, "Expected:", 200)
 	}
 
 	var ta taxiiAPIRoot
-	err := json.Unmarshal([]byte(result), &ta)
+	err = json.Unmarshal([]byte(result), &ta)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -27,7 +33,13 @@ func TestHandleTaxiiAPIRootFailRead(t *testing.T) {
 	renameFile("backend/sqlite/read/taxiiAPIRoot.sql", "backend/sqlite/read/taxiiAPIRoot.sql.testing")
 	defer renameFile("backend/sqlite/read/taxiiAPIRoot.sql.testing", "backend/sqlite/read/taxiiAPIRoot.sql")
 
-	status, _ := handlerTest(handleTaxiiAPIRoot, "GET", apiRootURL, nil)
+	ts, err := newTaxiiStorer(config.DataStore["name"], config.DataStore["path"])
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer ts.disconnect()
+
+	status, _ := handlerTest(handleTaxiiAPIRoot(ts), "GET", testAPIRootURL, nil)
 
 	if status != 400 {
 		t.Error("Got:", status, "Expected: 400")
@@ -37,14 +49,20 @@ func TestHandleTaxiiAPIRootFailRead(t *testing.T) {
 func TestHandleTaxiiAPIRootNotFound(t *testing.T) {
 	defer setupSQLite()
 
-	s, err := newSQLiteDB()
+	s, err := newSQLiteDB(config.DataStore["path"])
 	if err != nil {
 		t.Fatal(err)
 	}
 
 	_, err = s.db.Exec("delete from taxii_api_root")
 
-	status, _ := handlerTest(handleTaxiiAPIRoot, "GET", apiRootURL, nil)
+	ts, err := newTaxiiStorer(config.DataStore["name"], config.DataStore["path"])
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer ts.disconnect()
+
+	status, _ := handlerTest(handleTaxiiAPIRoot(ts), "GET", testAPIRootURL, nil)
 
 	if status != 404 {
 		t.Error("Got:", status, "Expected: 400")
