+++ test/config/main_test_config.json
@@ -1,10 +1,10 @@
 {
   "host": "localhost",
-  "port": 1234,
+  "port": 1235,
   "ssl_cert": "server.crt",
   "ssl_key": "server.key",
   "data_store": {
     "name": "sqlite",
-    "path": "db/cabby.db"
+    "path": "test/test.db"
   }
 }
